requirements.txt
agentscope==1.0.17
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
attrs==25.4.0
beniget==0.4.2.post1
bidict==0.23.1
cached-property==2.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
chromadb
cryptography==46.0.5
dashscope==1.25.13
distro==1.9.0
docstring_parser==0.17.0
flask==3.0.0
frozenlist==1.8.0
gast==0.6.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
jiter==0.13.0
json5==0.13.0
json_repair==0.58.5
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain-core==1.2.19
langchain-openai==1.1.11
deepagents==0.4.11
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.17
lightgbm==4.6.0
mcp==1.26.0
Mesa==3.5.0
multidict==6.7.1
networkx>=3.0
numpy==2.4.3
openai==2.28.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.1
ply==3.11
propcache==0.4.1
protobuf==6.33.5
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
python-datauri==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-frontmatter==1.1.0
python-multipart==0.0.22
python-socketio==5.16.1
pythran==0.18.1
pytz>=2024.1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
scikit-learn>=1.0.0
scipy==1.17.1
setuptools==82.0.1
shortuuid==1.0.13
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
SQLAlchemy==2.0.48
sse-starlette==3.3.2
starlette==0.52.1
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
websocket-client==1.9.0
wsproto==1.3.2
xgboost==3.2.0
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
